Remove Gemfile.lock from version control
The *general* best practice for application deployment is to commit
the Gemfile.lock to the repo for a reproducable environment between
developers and deployment. Since this site is a rarely maintained
github pages site, this leads to a need to occasionally update the
lock file due to vulnerable dependencies. Github Pages builds with the
latest jekyll version regardless and ignores the Gemfile.lock, so for
now remove it to ease the maintenance burden.